+++ agents/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @livekit/agents
 
+## 1.0.5
+
+### Patch Changes
+
+- add some logging to debug track not being subscribed - [#708](https://github.com/livekit/agents-js/pull/708) ([@Shubhrakanti](https://github.com/Shubhrakanti))
+
+## 1.0.4
+
+### Patch Changes
+
+- Add avatar datastream io component - [#687](https://github.com/livekit/agents-js/pull/687) ([@toubatbrian](https://github.com/toubatbrian))
+
+- bugfix: agent not recieving audio from room sometimes - [#705](https://github.com/livekit/agents-js/pull/705) ([@Shubhrakanti](https://github.com/Shubhrakanti))
+
+- refactor speech handle to allow waiting for playout in tool calls - [#693](https://github.com/livekit/agents-js/pull/693) ([@Shubhrakanti](https://github.com/Shubhrakanti))
+
 ## 1.0.3
 
 ### Patch Changes